requirements.txt
llama_cpp_python==0.3.8
rich==13.9.4
psutil==5.9.8
regex==2024.11.6
tqdm==4.67.1
numpy==2.2.4
pydantic==2.10.6
python-dotenv==1.0.1
jinja2==3.1.6
diskcache==5.6.3
typing_extensions==4.13.0 
pyopencl==2025.1